+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/Client.scala
@@ -352,9 +352,9 @@ trait Client
         * XPUBs, balances, etc). It's therefore not a good idea to enable
         * this logging in production.
         */
-      logger.info(
-        s"Command: $command ${parameters.map(_.toString).mkString(" ")}")
-      logger.info(s"Payload: \n${Json.prettyPrint(payload)}")
+//      logger.info(
+//        s"Command: $command ${parameters.map(_.toString).mkString(" ")}")
+//      logger.info(s"Payload: \n${Json.prettyPrint(payload)}")
       parseResult(result = (payload \ resultKey).validate[T],
                   json = payload,
                   printError = printError,
