+++ core/src/main/scala/org/bitcoins/core/p2p/NetworkPayload.scala
@@ -3,7 +3,7 @@ package org.bitcoins.core.p2p
 import org.bitcoins.core.bloom.{BloomFilter, BloomFlag}
 import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.gcs.{FilterHeader, FilterType, GolombFilter}
-import org.bitcoins.core.number._
+import org.bitcoins.core.number.{Int32, Int64, UInt16, UInt32, UInt64}
 import org.bitcoins.core.protocol.CompactSizeUInt
 import org.bitcoins.core.protocol.blockchain.{Block, BlockHeader, MerkleBlock}
 import org.bitcoins.core.protocol.transaction.Transaction
@@ -11,6 +11,7 @@ import org.bitcoins.core.serializers.p2p.messages._
 import org.bitcoins.core.util.BytesUtil
 import org.bitcoins.core.wallet.fee.{SatoshisPerByte, SatoshisPerKiloByte}
 import org.bitcoins.crypto._
+
 import scodec.bits.ByteVector
 
 /** Trait that represents a payload for a message on the Bitcoin p2p network
