+++ node-test/src/test/scala/org/bitcoins/node/NeutrinoNodeTest.scala
@@ -2,13 +2,15 @@ package org.bitcoins.node
 
 import akka.actor.Cancellable
 import org.bitcoins.asyncutil.AsyncUtil
+import org.bitcoins.core.util.NetworkUtil
 import org.bitcoins.crypto.DoubleSha256DigestBE
+import org.bitcoins.node.models.{Peer, PeerDAO, PeerDb}
 import org.bitcoins.server.BitcoinSAppConfig
 import org.bitcoins.testkit.BitcoinSTestAppConfig
-import org.bitcoins.testkit.node.{NodeTestUtil, NodeTestWithCachedBitcoindPair}
 import org.bitcoins.testkit.node.fixture.NeutrinoNodeConnectedWithBitcoinds
+import org.bitcoins.testkit.node.{NodeTestUtil, NodeTestWithCachedBitcoindPair}
 import org.bitcoins.testkit.util.TorUtil
-import org.scalatest.{FutureOutcome, Outcome}
+import org.scalatest.{Assertion, FutureOutcome, Outcome}
 
 import scala.concurrent.Future
 import scala.concurrent.duration.DurationInt
@@ -37,6 +39,104 @@ class NeutrinoNodeTest extends NodeTestWithCachedBitcoindPair {
   }
 
   behavior of "NeutrinoNode"
+
+  it must "be able to connect, initialize and then disconnect from all peers" in {
+    nodeConnectedWithBitcoind: NeutrinoNodeConnectedWithBitcoinds =>
+      val node = nodeConnectedWithBitcoind.node
+      def peerManager = node.peerManager
+      def peers = peerManager.peers
+      val ourPeersF: Future[Vector[Peer]] = Future.sequence(
+        nodeConnectedWithBitcoind.bitcoinds.map(NodeTestUtil.getBitcoindPeer))
+
+      def has2Peers: Future[Unit] =
+        AsyncUtil.retryUntilSatisfied(peers.size == 2,
+                                      interval = 1.second,
+                                      maxTries = 5)
+      def bothOurs: Future[Assertion] = ourPeersF.map { ours =>
+        assert(ours.map(peers.contains(_)).forall(_ == true))
+      }
+      def allConnected: Future[Assertion] = for {
+        conns <- Future.sequence(peers.map(peerManager.isConnected))
+      } yield assert(conns.forall(_ == true))
+      def allInitialized: Future[Assertion] = for {
+        inits <- Future.sequence(peers.map(peerManager.isInitialized))
+      } yield assert(inits.forall(_ == true))
+
+      def allDisconn: Future[Unit] = AsyncUtil.retryUntilSatisfied(
+        peers
+          .map(p =>
+            !peerManager.peerData.contains(p) && !peerManager.waitingForDeletion
+              .contains(p))
+          .forall(_ == true),
+        maxTries = 5,
+        interval = 1.second
+      )
+
+      for {
+        _ <- has2Peers
+        _ <- bothOurs
+        _ <- allConnected
+        _ <- allInitialized
+        _ = peers.map(peerManager.removePeer)
+        _ <- allDisconn
+      } yield {
+        succeed
+      }
+  }
+
+  it must "store peers after successful initialization" in {
+    nodeConnectedWithBitcoind: NeutrinoNodeConnectedWithBitcoinds =>
+      val node = nodeConnectedWithBitcoind.node
+      val peerManager = node.peerManager
+      def peers = peerManager.peers
+
+      //is the database peer same as Peer
+      def isSame(peerDb: PeerDb, peer: Peer): Boolean = {
+        val dbSocket =
+          NetworkUtil.parseInetSocketAddress(peerDb.address, peerDb.port)
+
+        val hostMatch: Boolean = {
+          if (dbSocket.getHostString == peer.socket.getHostString) true
+          else {
+            //checking if both are localhost
+            //a bit hacky way but resolution of localhost to address cannot be done so as to allow for tor
+            //addresses too
+            val localhost = Vector("localhost", "127.0.0.1")
+            localhost.contains(dbSocket.getHostString) && localhost.contains(
+              peer.socket.getHostString)
+          }
+        }
+
+        hostMatch && dbSocket.getPort == peer.socket.getPort
+      }
+
+      //assert connected to 2 peers and both initialised and connected
+      val assertConnAndInit = for {
+        _ <- AsyncUtil
+          .retryUntilSatisfied(peers.size == 2,
+                               interval = 1.second,
+                               maxTries = 5)
+        _ <- Future
+          .sequence(peers.map(peerManager.isConnected))
+          .flatMap(p => assert(p.forall(_ == true)))
+        res <- Future
+          .sequence(peers.map(peerManager.isConnected))
+          .flatMap(p => assert(p.forall(_ == true)))
+      } yield res
+
+      for {
+        _ <- assertConnAndInit
+        ourPeers <- Future.sequence(
+          nodeConnectedWithBitcoind.bitcoinds.map(NodeTestUtil.getBitcoindPeer))
+        peerDbs <- PeerDAO()(executionContext, node.nodeAppConfig).findAll()
+      } yield {
+
+        val allInDb = ourPeers.forall { p =>
+          peerDbs.exists(isSame(_, p))
+        }
+        assert(allInDb)
+      }
+  }
 
   it must "receive notification that a block occurred on the p2p network for neutrino" in {
     nodeConnectedWithBitcoind: NeutrinoNodeConnectedWithBitcoinds =>
